+++ carp_webservices/test/carp_service_test.dart
@@ -7,8 +7,7 @@ import 'package:carp_webservices/carp_auth/carp_auth.dart';
 import 'package:carp_webservices/carp_service/carp_service.dart';
 import 'package:test/test.dart';
 
-String _encode(Object object) =>
-    const JsonEncoder.withIndent(' ').convert(object);
+String _encode(Object object) => const JsonEncoder.withIndent(' ').convert(object);
 
 void main() {
   //final String username = "researcher@example.com";
@@ -67,8 +66,7 @@ void main() {
     setUp(() {});
 
     test('- authentication', () async {
-      user = await CarpService.instance
-          .authenticate(username: username, password: password);
+      user = await CarpService.instance.authenticate(username: username, password: password);
 
       assert(user != null);
       assert(user.token != null);
@@ -90,10 +88,8 @@ void main() {
     // This test fails -- we do not have access to create users with the authenticated user.
     test('- create user', () async {
       int id = random.nextInt(1000);
-      CarpUser new_user = await CarpService.instance.createUser(
-          username: 'user_$id@dtu.dk',
-          password: 'underbar',
-          firstName: 'CACHET User #$id');
+      CarpUser new_user = await CarpService.instance
+          .createUser(username: 'user_$id@dtu.dk', password: 'underbar', firstName: 'CACHET User #$id');
 
       // we expect this call to fail, since we're not authenticated as admin
       assert(new_user == null);
@@ -140,8 +136,7 @@ void main() {
     });
 
     test('- authentication with saved token', () async {
-      CarpUser new_user = await CarpService.instance
-          .authenticateWithToken(username: username, token: user.token);
+      CarpUser new_user = await CarpService.instance.authenticateWithToken(username: username, token: user.token);
 
       assert(new_user != null);
       assert(new_user.isAuthenticated);
@@ -151,8 +146,8 @@ void main() {
     });
 
     test('- authentication with saved JSON token', () async {
-      CarpUser new_user = await CarpService.instance.authenticateWithToken(
-          username: username, token: OAuthToken.fromJson(tokenAsJson));
+      CarpUser new_user =
+          await CarpService.instance.authenticateWithToken(username: username, token: OAuthToken.fromJson(tokenAsJson));
 
       assert(new_user != null);
       assert(new_user.isAuthenticated);
@@ -167,10 +162,7 @@ void main() {
     () {
       test('- create', () async {
         ConsentDocument uploaded = await CarpService.instance
-            .createConsentDocument({
-          "text": "The original terms text.",
-          "signature": "Image Blob"
-        });
+            .createConsentDocument({"text": "The original terms text.", "signature": "Image Blob"});
 
         assert(uploaded != null);
         print(uploaded);
@@ -180,29 +172,25 @@ void main() {
       });
 
       test('- get', () async {
-        ConsentDocument uploaded =
-            await CarpService.instance.getConsentDocument(consentDocumentId);
+        ConsentDocument uploaded = await CarpService.instance.getConsentDocument(consentDocumentId);
 
         assert(uploaded != null);
         print(uploaded);
         print(uploaded.createdAt);
       });
     },
-    skip: true,
+    skip: false,
   );
 
   group(
     "Data points",
     () {
       test('- post', () async {
-        final CARPDataPoint data =
-            CARPDataPoint.fromDatum(study.id, study.userId, datum);
+        final CARPDataPoint data = CARPDataPoint.fromDatum(study.id, study.userId, datum);
 
         print(_encode(data.toJson()));
 
-        dataPointId = await CarpService.instance
-            .getDataPointReference()
-            .postDataPoint(data);
+        dataPointId = await CarpService.instance.getDataPointReference().postDataPoint(data);
 
         assert(dataPointId > 0);
         print("data_point_id : $dataPointId");
@@ -210,16 +198,12 @@ void main() {
 
       test('- batch', () async {
         final File file = File("test/batch.json");
-        await CarpService.instance
-            .getDataPointReference()
-            .batchPostDataPoint(file);
+        await CarpService.instance.getDataPointReference().batchPostDataPoint(file);
       });
 
       test('- get by id', () async {
         print("GET data_point_id : $dataPointId");
-        CARPDataPoint data = await CarpService.instance
-            .getDataPointReference()
-            .getDataPoint(dataPointId);
+        CARPDataPoint data = await CarpService.instance.getDataPointReference().getDataPoint(dataPointId);
 
         print(_encode(data.toJson()));
         assert(data.id == dataPointId);
@@ -227,22 +211,17 @@ void main() {
       });
 
       test('- get all', () async {
-        List<CARPDataPoint> data = await CarpService.instance
-            .getDataPointReference()
-            .getAllDataPoint();
+        List<CARPDataPoint> data = await CarpService.instance.getDataPointReference().getAllDataPoint();
 
         data.forEach((datapoint) => print(_encode((datapoint.toJson()))));
         assert(data.length > 0);
       });
 
       test('- query', () async {
-        String query =
-            'carp_header.user_id==$userId;carp_body.timestamp>2019-11-02T12:53:40.219598Z';
+        String query = 'carp_header.user_id==$userId;carp_body.timestamp>2019-11-02T12:53:40.219598Z';
         //String query = 'carp_header.user_id==$userId';
         print("query : $query");
-        List<CARPDataPoint> data = await CarpService.instance
-            .getDataPointReference()
-            .queryDataPoint(query);
+        List<CARPDataPoint> data = await CarpService.instance.getDataPointReference().queryDataPoint(query);
 
         data.forEach((datapoint) => print(_encode((datapoint.toJson()))));
         assert(data.length > 0);
@@ -250,9 +229,7 @@ void main() {
 
       test('- delete', () async {
         print("DELETE data_point_id : $dataPointId");
-        await CarpService.instance
-            .getDataPointReference()
-            .deleteDataPoint(dataPointId);
+        await CarpService.instance.getDataPointReference().deleteDataPoint(dataPointId);
       });
     },
     skip: false,
@@ -260,7 +237,7 @@ void main() {
 
   group("Documents & Collections", () {
     test(' - add document', () async {
-      // is not providing an document id, so this should create a new document
+      // is providing userId as the document name
       // if the collections (patients) don't exist, it is created (according to David).
       document = await CarpService.instance
           .collection('patients')
@@ -284,10 +261,7 @@ void main() {
       assert(document != null);
       print(document);
 
-      DocumentSnapshot original = await CarpService.instance
-          .collection('patients')
-          .document(document.name)
-          .get();
+      DocumentSnapshot original = await CarpService.instance.collection('patients').document(document.name).get();
       print(_encode(original.data));
 
       // updating the role to super user
@@ -306,8 +280,7 @@ void main() {
 
     test(' - get document by id', () async {
       assert(document != null);
-      DocumentSnapshot newDocument =
-          await CarpService.instance.documentById(documentId).get();
+      DocumentSnapshot newDocument = await CarpService.instance.documentById(documentId).get();
 
       print((newDocument));
       assert(newDocument.id == document.id);
@@ -315,10 +288,7 @@ void main() {
     });
 
     test(' - get document by path', () async {
-      DocumentSnapshot newDocument = await CarpService.instance
-          .collection('patients')
-          .document(document.name)
-          .get();
+      DocumentSnapshot newDocument = await CarpService.instance.collection('patients').document(document.name).get();
       print((newDocument));
       assert(newDocument.id == document.id);
     });
@@ -331,10 +301,8 @@ void main() {
       print(_encode(document.data));
 
       print('----------- renamed document -------------');
-      DocumentSnapshot renamed_document = await CarpService.instance
-          .collection('patients')
-          .document(document.name)
-          .rename('new_name');
+      DocumentSnapshot renamed_document =
+          await CarpService.instance.collection('patients').document(document.name).rename('new_name');
       print(renamed_document);
       print(_encode(renamed_document.data));
 
@@ -343,16 +311,12 @@ void main() {
 //          await CarpService.instance.collection('patients').document(renamed_document.name).get();
 
       print('----------- server document by ID -------------');
-      DocumentSnapshot server_document =
-          await CarpService.instance.documentById(documentId).get();
+      DocumentSnapshot server_document = await CarpService.instance.documentById(documentId).get();
       print(server_document);
       print(_encode(server_document.data));
 
       print('----------- server document by NAME -------------');
-      server_document = await CarpService.instance
-          .collection('patients')
-          .document(renamed_document.name)
-          .get();
+      server_document = await CarpService.instance.collection('patients').document(renamed_document.name).get();
       print(server_document);
       print(_encode(server_document.data));
 
@@ -364,8 +328,7 @@ void main() {
     test(' - get document by query', () async {
       assert(document != null);
       String query = 'name==$userId';
-      List<DocumentSnapshot> documents =
-          await CarpService.instance.documentsByQuery(query);
+      List<DocumentSnapshot> documents = await CarpService.instance.documentsByQuery(query);
 
       print('Found ${documents.length} document(s)');
       documents.forEach((document) => print(' - $document'));
@@ -420,35 +383,30 @@ void main() {
     });
 
     test(" - list documents in 'patients' collection", () async {
-      List<DocumentSnapshot> documents =
-          await CarpService.instance.collection("patients").documents;
+      List<DocumentSnapshot> documents = await CarpService.instance.collection("patients").documents;
       documents.forEach((doc) => print(doc));
     });
 
     test(" - list collections in the 'user@dtu.dk' document", () async {
-      DocumentSnapshot newDocument = await CarpService.instance
-          .collection('patients')
-          .document(userId)
-          .get();
+      DocumentSnapshot newDocument = await CarpService.instance.collection('patients').document(userId).get();
       newDocument.collections.forEach((ref) => print(ref));
     });
 
     test(" - list all nested documents in 'patients' collection", () async {
-      List<DocumentSnapshot> documents =
-          await CarpService.instance.collection("patients").documents;
+      List<DocumentSnapshot> documents = await CarpService.instance.collection("patients").documents;
       documents.forEach((doc) {
         print(doc);
         doc.collections.forEach((col) => print(col));
       });
+    });
 
-      //     List<DocumentSnapshot> documents;
-
-//      // List all collections in the root
+//    test(" - list all collections in the root", () async {
 //      List<String> root = await CarpService.instance.collection("").collections;
 //      for (String ref in root) {
 //        print(ref);
 //        // List all documents in each collection
-//        documents = await CarpService.instance.collection("/$ref").documents;
+//        List<DocumentSnapshot> documents =
+//            await CarpService.instance.collection("/$ref").documents;
 //        for (DocumentSnapshot doc in documents) {
 //          print(doc);
 //        }
@@ -458,34 +416,28 @@ void main() {
 //      for (DocumentSnapshot doc in documents) {
 //        print(doc);
 //      }
-    });
+//    });
 
     test(' - get collection from path', () async {
-      CollectionReference collection =
-          await CarpService.instance.collection('patients').get();
+      CollectionReference collection = await CarpService.instance.collection('patients').get();
       assert(collection.id > 0);
       print(collection);
     });
 
     test(' - delete document', () async {
       assert(document != null);
-      await CarpService.instance
-          .collection('patients')
-          .document(document.name)
-          .delete();
+      await CarpService.instance.collection('patients').document(document.name).delete();
     });
 
     test(' - rename collection', () async {
-      CollectionReference collection =
-          await CarpService.instance.collection('patients').get();
+      CollectionReference collection = await CarpService.instance.collection('patients').get();
       await collection.rename('new_patients');
       expect(collection.name, 'new_patients');
       print(collection);
     });
 
     test(' - delete collection', () async {
-      CollectionReference collection =
-          await CarpService.instance.collection('new_patients').get();
+      CollectionReference collection = await CarpService.instance.collection('new_patients').get();
       await collection.delete();
       print(collection);
     });
@@ -501,11 +453,7 @@ void main() {
 
         final FileUploadTask uploadTask = CarpService.instance
             .getFileStorageReference()
-            .upload(myFile, {
-          'content-type': 'image/jpg',
-          'content-language': 'en',
-          'activity': 'test'
-        });
+            .upload(myFile, {'content-type': 'image/jpg', 'content-language': 'en', 'activity': 'test'});
 
         assert(uploadTask != null);
 
@@ -519,8 +467,7 @@ void main() {
       });
 
       test('- get', () async {
-        final CarpFileResponse result =
-            await CarpService.instance.getFileStorageReference(id).get();
+        final CarpFileResponse result = await CarpService.instance.getFileStorageReference(id).get();
 
         assert(result.id == id);
         print('result : $result');
@@ -529,8 +476,7 @@ void main() {
       test('- download', () async {
         final File myFile = File("test/img-$id.jpg");
 
-        final FileDownloadTask downloadTask =
-            CarpService.instance.getFileStorageReference(id).download(myFile);
+        final FileDownloadTask downloadTask = CarpService.instance.getFileStorageReference(id).download(myFile);
 
         assert(downloadTask != null);
 
@@ -540,21 +486,19 @@ void main() {
       });
 
       test('- get all', () async {
-        final List<CarpFileResponse> results =
-            await CarpService.instance.getFileStorageReference(id).getAll();
+        final List<CarpFileResponse> results = await CarpService.instance.getFileStorageReference(id).getAll();
 
         //assert(result.id == id);
         print('result : $results');
       });
 
       test('- delete', () async {
-        final int result =
-            await CarpService.instance.getFileStorageReference(id).delete();
+        final int result = await CarpService.instance.getFileStorageReference(id).delete();
 
         assert(result > 0);
         print('result : $result');
       });
     },
-    skip: true,
+    skip: false,
   );
 }