+++ carp_webservices/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.4.4
+* DataPoint and Informed Consent endpoint changed to use `deployments` in URL.
+* Aligned the collection/documents to the new CANS endpoints
+
 ## 0.4.3
 * `study_id` is now a String (not an int anymore).
 * fixed issue [#66](https://github.com/cph-cachet/carp.sensing-flutter/issues/66).